+++ test/unit/gen_class/argparser_arg_set_path.py
@@ -34,6 +34,10 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_no_path_no_cmd
+        test_no_path_cmd
+        test_path_no_cmd
+        test_path_cmd
         test_trailing_slash
         test_arg_not_present
         test_arg_present
@@ -54,20 +58,84 @@ class UnitTest(unittest.TestCase):
         """
 
         p_name = "program.py"
-
         self.argv = [p_name]
         self.argv2 = [p_name, "-a", "/dir/path", "-c", "/dir/diff_path"]
         self.argv3 = [p_name, "-a", "/dir/path/"]
-
         self.opt_val = ["-a", "-c"]
-
         self.arg_opt = None
         self.arg_opt2 = "-a"
         self.arg_opt3 = "-b"
-
+        self.cmd = "mysql"
         self.results = ""
-        self.results2 = "/dir/path"
-        self.results3 = "/dir/path/"
+        self.results2 = "/dir/path/"
+        self.results3 = "/dir/path/mysql"
+        self.results4 = "mysql"
+
+    def test_no_path_no_cmd(self):
+
+        """Function:  test_no_path_no_cmd
+
+        Description:  Test with path, but no command passed.
+
+        Arguments:
+
+        """
+
+        args_array = gen_class.ArgParser(
+            self.argv2, opt_val=self.opt_val, do_parse=True)
+
+        self.assertEqual(
+            args_array.arg_set_path(self.arg_opt3), self.results)
+
+    def test_no_path_cmd(self):
+
+        """Function:  test_no_path_cmd
+
+        Description:  Test with no path, but command passed.
+
+        Arguments:
+
+        """
+
+        args_array = gen_class.ArgParser(
+            self.argv2, opt_val=self.opt_val, do_parse=True)
+
+        self.assertEqual(
+            args_array.arg_set_path(
+                self.arg_opt3, cmd=self.cmd), self.results4)
+
+    def test_path_no_cmd(self):
+
+        """Function:  test_path_no_cmd
+
+        Description:  Test with path, but no command passed.
+
+        Arguments:
+
+        """
+
+        args_array = gen_class.ArgParser(
+            self.argv2, opt_val=self.opt_val, do_parse=True)
+
+        self.assertEqual(
+            args_array.arg_set_path(self.arg_opt2), self.results2)
+
+    def test_path_cmd(self):
+
+        """Function:  test_path_cmd
+
+        Description:  Test with path and command passed.
+
+        Arguments:
+
+        """
+
+        args_array = gen_class.ArgParser(
+            self.argv2, opt_val=self.opt_val, do_parse=True)
+
+        self.assertEqual(
+            args_array.arg_set_path(
+                self.arg_opt2, cmd=self.cmd), self.results3)
 
     def test_trailing_slash(self):
 
@@ -82,7 +150,7 @@ class UnitTest(unittest.TestCase):
         args_array = gen_class.ArgParser(
             self.argv3, opt_val=self.opt_val, do_parse=True)
 
-        self.assertEqual(args_array.arg_set_path(self.arg_opt2), self.results3)
+        self.assertEqual(args_array.arg_set_path(self.arg_opt2), self.results2)
 
     def test_arg_not_present(self):